requirements.txt
dnspython>=2.6.1
requests>=2.25.1
lxml>=4.6.3
asyncio>=3.4.3